agentic_rag/components/evaluators/reference_free_evaluator.py
"""Reference-free evaluator component (no gold standard needed)."""

from typing import Any, Dict, List, Optional

from haystack import component, default_from_dict, default_to_dict


@component
class ReferenceFreeEvaluator:
    """
    Evaluates retrieval and generation quality without gold standard.

    Computes:
    - Faithfulness: Does answer align with retrieved documents?
    - Context Relevance: Are retrieved documents relevant to query?

    Takes pipeline outputs, adds eval_metrics, passes everything through.
    """

    def __init__(self) -> None:
        """Initialize the reference-free evaluator."""
        pass

    @component.output_types(eval_data=Dict[str, Any])  # type: ignore[misc]
    def run(
        self,
        query: str,
        replies: Optional[List[str]] = None,
        eval_data: Optional[Dict[str, Any]] = None,
        ground_truth_answer: Optional[str] = None,
        relevant_doc_ids: Optional[List[str]] = None,
    ) -> Dict[str, Any]:
        """
        Run reference-free evaluation on answer quality.

        Args:
            query: User query
            replies: Generated answers
            eval_data: Evaluation dict (created if None, updated if exists)
            ground_truth_answer: Passed through (not used by this evaluator)
            relevant_doc_ids: Passed through (not used by this evaluator)

        Returns:
            Dict with single key 'eval_data' containing all results
        """
        # Initialize or update eval_data
        if eval_data is None:
            eval_data = {
                "query": query,
                "answer": replies[0] if replies else None,
                "ground_truth_answer": ground_truth_answer,
                "relevant_doc_ids": relevant_doc_ids,
                "eval_metrics": {},
            }

        # Update with current data if not already set
        if "query" not in eval_data:
            eval_data["query"] = query
        if "answer" not in eval_data and replies:
            eval_data["answer"] = replies[0]
        if "ground_truth_answer" not in eval_data:
            eval_data["ground_truth_answer"] = ground_truth_answer
        if "relevant_doc_ids" not in eval_data:
            eval_data["relevant_doc_ids"] = relevant_doc_ids

        # Ensure eval_metrics exists
        if "eval_metrics" not in eval_data:
            eval_data["eval_metrics"] = {}

        answer = replies[0] if replies else eval_data.get("answer")

        # Compute reference-free metrics (answer quality only)
        metrics = {}

        # TODO: Implement actual evaluation using Haystack evaluators
        # For now, placeholder metrics
        if answer and query:
            # Answer completeness: Does answer address the query?
            metrics["answer_completeness"] = {
                "score": 0.88,  # Placeholder - check if answer is complete
                "type": "reference_free",
            }

            # Answer clarity: Is answer clear and well-formed?
            metrics["answer_clarity"] = {
                "score": 0.91,  # Placeholder - check grammar, coherence
                "type": "reference_free",
            }

        # Add metrics to eval_data
        eval_data["eval_metrics"].update(metrics)

        # Return single dict with everything inside
        return {"eval_data": eval_data}

    def to_dict(self) -> Dict[str, Any]:
        """Serialize component to dict."""
        return default_to_dict(self)  # type: ignore[no-any-return]

    @classmethod
    def from_dict(cls, data: Dict[str, Any]) -> "ReferenceFreeEvaluator":
        """Deserialize component from dict."""
        return default_from_dict(cls, data)  # type: ignore[no-any-return]
