pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = ""
authors = [
    {name = "Vardhan Shorewala",email = "vardhanshorewala@berkeley.edu"}
]
# readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "haystack-ai>=2.0.0",
    # Required for PDF conversion
    "pypdf>=3.0.0",
    # Required for document splitting/chunking
    "nltk>=3.9.1",
    # Required for embeddings
    "sentence-transformers>=2.0.0",
    # Required for Chroma document store
    "chroma-haystack>=0.16.0",
    # Required for OpenAI components
    "openai>=1.0.0",
    # Required for Marker PDF conversion
    "marker-pdf>=1.9.3",
    "markitdown[pdf] (>=0.1.3,<0.2.0)",
    "reportlab (>=4.4.3,<5.0.0)",
    # Required for Neo4j graph database
    "neo4j (>=5.28.0,<5.29.0)",
    "certifi>=2023.0.0",
    "dotenv (>=0.9.9,<0.10.0)",
    "fpdf (>=1.7.2,<2.0.0)",
    # Fix for web3/eth_typing compatibility issue
    "web3>=7.0.0",
    "eth-typing>=5.0.0",
    "rouge-score (>=0.1.2,<0.2.0)",
    "scipy (>=1.16.3,<2.0.0)",
    "types-requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.0.0",
    "pytest>=7.0.0",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

# Black configuration
[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["agentic_rag"]

# flake8 configuration (in setup.cfg since flake8 doesn't support pyproject.toml natively)
# See .flake8 file for configuration

# mypy configuration
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
